--- a/.dockerignore
+++ b/.dockerignore
@@ -1,0 +1,58 @@
+# Dependencies
+node_modules
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Next.js
+.next
+out
+
+# Testing
+coverage
+.jest
+
+# Misc
+.DS_Store
+*.pem
+
+# Environment files
+.env
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# Git
+.git
+.gitignore
+
+# IDE
+.vscode
+.idea
+*.swp
+*.swo
+*.swn
+.cursor
+
+# Documentation
+README.md
+docs
+*.md
+
+# CI/CD
+.github
+.husky
+
+# Infrastructure
+infra
+
+# Playwright
+.playwright-mcp
+tests
+playwright-report
+playwright.config.ts
+
+# Build artifacts
+dist
+build