## Summary

Implements complete ECS Fargate infrastructure to replace AWS Amplify hosting, enabling native HTTP/2 streaming support for AI chat responses with improved scalability and observability.

**Part of Epic #305**

## Changes

### Infrastructure
- ✅ New ECS Fargate service with Application Load Balancer
- ✅ Multi-stage Dockerfile optimized for Next.js standalone builds
- ✅ Auto-scaling configuration (1-10 containers based on CPU/memory)
- ✅ AWS WAF with rate limiting and managed rule sets
- ✅ CloudWatch dashboard for comprehensive monitoring
- ✅ SSL/TLS termination at ALB with ACM certificates
- ✅ VPC sharing with DatabaseStack for cost optimization

### Components Created
- `/Dockerfile` - Multi-stage container build for Next.js
- `/.dockerignore` - Optimized build context exclusions
- `/infra/lib/constructs/ecs-service.ts` - Reusable ECS service construct
- `/infra/lib/frontend-stack-ecs.ts` - Complete deployment stack
- `/infra/bin/infra.ts` - Updated to deploy new ECS stacks

### Key Features
- **HTTP/2 Streaming**: Native support for Server-Sent Events
- **Auto-scaling**: CPU (70%) and memory (80%) based scaling
- **Container Insights**: Detailed metrics and performance monitoring
- **Fargate Spot**: 70% cost savings in dev environment
- **ARM64 Architecture**: 20% cost savings with Graviton2
- **Security**: Non-root user, WAF protection, ECR scanning

## Architecture

```mermaid
graph TB
    Users[Users] -->|HTTPS| ALB[Application Load Balancer]
    ALB -->|HTTP| ECS[ECS Fargate Tasks]
    ECS --> RDS[Aurora Serverless v2]
    ECS --> S3[S3 Documents]
    ECS --> Bedrock[AWS Bedrock]
    WAF[AWS WAF] --> ALB
    CI[Container Insights] --> CW[CloudWatch]
    ECS --> CW
```

## Testing

- [x] CDK synth validation passed
- [x] TypeScript compilation successful
- [x] ESLint validation passed
- [x] Docker build tested locally
- [x] Health check endpoint validated
- [x] IAM permissions match current SSR Compute Role

## Deployment Strategy

**Blue-Green Approach:**

1. Deploy ECS stack alongside existing Amplify
2. Build and push Docker image to ECR
3. Configure environment variables
4. Gradual traffic shift via Route53 weighted routing
5. Full cutover after validation
6. Decommission Amplify (code retained for rollback)

## Deployment Commands

```bash
# Dev environment
cd infra
npx cdk deploy AIStudio-FrontendStack-ECS-Dev --context baseDomain=aistudio.psd401.ai

# Prod environment
npx cdk deploy AIStudio-FrontendStack-ECS-Prod --context baseDomain=aistudio.psd401.ai
```

## Environment Variables

Same as current Amplify configuration:
- `AUTH_*` - NextAuth configuration
- `NEXT_PUBLIC_COGNITO_*` - Cognito settings
- `RDS_*` - Database connection
- `AWS_REGION` - Automatically provided by ECS

## Cost Impact

| Environment | Current (Amplify) | New (ECS) | Delta |
|-------------|-------------------|-----------|-------|
| Dev | ~$40/mo | ~$35/mo | -12% (Fargate Spot) |
| Prod | ~$150/mo | ~$175/mo | +17% |

Total increase: ≤20% for significantly improved streaming and control.

## Monitoring

Automatic CloudWatch dashboard includes:
- ECS service CPU/memory utilization
- ALB request count and latency (P50, P95, P99)
- Task count and auto-scaling events
- Container-level metrics via Container Insights

Logs: `/ecs/aistudio-{dev|prod}`

## Security

- Non-root container execution (UID 1001)
- Security groups with least-privilege
- WAF rate limiting (2000 req/5min per IP)
- AWS Managed WAF rules (Core, SQLi, Bad Inputs)
- ECR vulnerability scanning enabled
- IAM roles scoped to necessary resources

## Breaking Changes

None. This is additive infrastructure.

## Rollback Plan

If issues arise:
1. Route53 immediate failback to Amplify (60s TTL)
2. ECS service scaled to 0 desired tasks
3. Investigate using CloudWatch logs
4. Restore Amplify stack from commented code if needed

## Next Steps

- [ ] Deploy to dev environment
- [ ] Build and push Docker image
- [ ] Configure environment variables
- [ ] Validate streaming functionality
- [ ] Monitor metrics for 48 hours
- [ ] Gradual traffic migration
- [ ] Production deployment
- [ ] Decommission Amplify after 1 week stable

## Checklist

- [x] Code follows project conventions
- [x] No TypeScript errors
- [x] Linting passed
- [x] Infrastructure validated with CDK synth
- [x] Documentation updated (deployment instructions in outputs)
- [x] Security considerations addressed
- [x] Monitoring and observability configured

Closes #306